+++ matrix-sdk/src/main/java/org/matrix/androidsdk/rest/client/ProfileRestClient.java
@@ -243,6 +243,7 @@ public class ProfileRestClient extends RestClient<ProfileApi> {
             forgetPasswordParams.client_secret = pid.clientSecret;
             forgetPasswordParams.send_attempt = 1;
 
+            // TODO privacy: there is something to do here, we cannot use identity server from hsConfig
             Uri identityServerUri = mHsConfig.getIdentityServerUri();
             if (identityServerUri == null) {
                 callback.onUnexpectedError(new IdentityServerNotConfiguredException());
@@ -369,6 +370,7 @@ public class ProfileRestClient extends RestClient<ProfileApi> {
         params.clientSecret = clientSecret;
         params.sendAttempt = attempt;
 
+        // TODO privacy: there is something to do here, we cannot use identity server from hsConfig
         Uri identityServerUri = mHsConfig.getIdentityServerUri();
         if (identityServerUri == null) {
             callback.onUnexpectedError(new IdentityServerNotConfiguredException());
@@ -429,6 +431,7 @@ public class ProfileRestClient extends RestClient<ProfileApi> {
         params.clientSecret = clientSecret;
         params.sendAttempt = attempt;
 
+        // TODO privacy: there is something to do here, we cannot use identity server from hsConfig
         Uri identityServerUri = mHsConfig.getIdentityServerUri();
         if (identityServerUri == null) {
             callback.onUnexpectedError(new IdentityServerNotConfiguredException());
@@ -477,6 +480,7 @@ public class ProfileRestClient extends RestClient<ProfileApi> {
         AddThreePidsParams params = new AddThreePidsParams();
         params.three_pid_creds = new ThreePidCreds();
 
+        // TODO privacy: there is something to do here, we cannot use identity server from hsConfig
         Uri identityServerUri = mHsConfig.getIdentityServerUri();
         if (identityServerUri == null) {
             callback.onUnexpectedError(new IdentityServerNotConfiguredException());