+++ CHANGES.rst
@@ -8,6 +8,7 @@ Improvements:
  - Display correctly the revoked third-party invites.
  - Support optional default STUN server when no ICE provided by HS
  - Use wellknown to discover the IS of a HS (vector-im/riot-android#3283)
+ - Make identity server configurable
 
 Bugfix:
  -